// src/grid/grid-classes.d.ts
export declare const WIDTHS: {
    [key: string]: string;
};
export declare const VIEW_WIDTHS: {
    [key: string]: object;
};
export declare const calculateGap: (gap: number, columnSize: number) => number;
export declare const populateGridStyle: (target: string, styles: any, gap: number, columnSize: number) => any;
export declare const CLASSES: {
    [key: string]: string;
    column: any;
    disableWrapper: any;
    disableWrapperActive: any;
    row: any;
};
export declare const CONTAINER_TYPE: {
    row: string;
    column: string;
};
export declare const ALIGN: {
    [key: string]: string;
};